Stuff shoots

+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/data/Entity.java
@@ -1,7 +1,7 @@
 package dk.sdu.mmmi.cbse.common.data;
 
 import dk.sdu.mmmi.cbse.common.data.entityparts.EntityPart;
-import java.awt.Color;
+import com.badlogic.gdx.graphics.Color;
 import java.io.Serializable;
 import java.util.Map;
 import java.util.UUID;
@@ -13,7 +13,7 @@ public class Entity implements Serializable {
     private float[] shapeX = new float[4];
     private float[] shapeY = new float[4];
     private float radius;
-    private Color colour = new Color(1,1,1,1);
+    private Color colour = new Color(1, 1, 1, 1);
     private Map<Class, EntityPart> parts;
     
     public Entity() {
@@ -64,7 +64,7 @@ public class Entity implements Serializable {
         return colour;
     }
     
-    public void setColour(int r, int g, int b, int a){
+    public void setColour(float r, float g, float b, float a){
         colour = new Color(r, g, b, a);
     }
 }
